Add detailed description of field notes.
For field notes, do some logic inside the template.

It might be useful to put that logic into the database model (or a
method of its) in the long run – I'll raise an issue about that – but
for the moment, the logic is at least in place somewhere.

Also remove RIS and MODS export, which don't work.